backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
bcrypt<4.0.0
passlib[bcrypt]
python-jose[cryptography]
python-multipart
Pillow
